# docs/api-specification.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Serverless E-Commerce API",
    "description": "RESTful API for a serverless e-commerce platform built on AWS",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "https://api.example.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://api-dev.example.com/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List all products",
        "description": "Retrieve a paginated list of products with optional filtering by category, price range, and search query",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by product category",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Electronics"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "description": "Minimum price filter",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float",
              "minimum": 0,
              "example": 10.0
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "description": "Maximum price filter",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float",
              "minimum": 0,
              "example": 500.0
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search query for product name or description",
            "required": false,
            "schema": {
              "type": "string",
              "example": "wireless"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20
            }
          },
          {
            "name": "nextToken",
            "in": "query",
            "description": "Pagination token from previous response",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create a new product (Admin only)",
        "description": "Create a new product in the catalog. Requires admin authentication.",
        "operationId": "createProduct",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication endpoint"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "required": ["productId", "name", "price", "currency", "category", "inventory", "status"],
        "properties": {
          "productId": {
            "type": "string",
            "description": "Unique product identifier",
            "example": "prod-123"
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Wireless Headphones"
          },
          "description": {
            "type": "string",
            "description": "Detailed product description",
            "example": "Premium noise-cancelling wireless headphones with 30-hour battery life"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Product price",
            "example": 199.99
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "USD"
          },
          "category": {
            "type": "string",
            "description": "Product category",
            "example": "Electronics"
          },
          "subCategory": {
            "type": "string",
            "description": "Product subcategory",
            "example": "Audio"
          },
          "brand": {
            "type": "string",
            "description": "Product brand",
            "example": "TechBrand"
          },
          "images": {
            "type": "array",
            "description": "Product image URLs",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "example": ["https://cdn.example.com/img1.jpg", "https://cdn.example.com/img2.jpg"]
          },
          "inventory": {
            "type": "integer",
            "description": "Available stock quantity",
            "example": 50
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping unit",
            "example": "WH-123-BLK"
          },
          "attributes": {
            "type": "object",
            "description": "Product-specific attributes",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "color": "Black",
              "weight": "250g",
              "batteryLife": "30 hours"
            }
          },
          "ratings": {
            "type": "object",
            "properties": {
              "average": {
                "type": "number",
                "format": "float",
                "example": 4.5
              },
              "count": {
                "type": "integer",
                "example": 128
              }
            }
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "discontinued"],
            "example": "active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-15T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-17T08:20:00Z"
          }
        }
      },
      "ProductListResponse": {
        "type": "object",
        "required": ["products", "count"],
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "count": {
            "type": "integer",
            "description": "Number of products in current response",
            "example": 20
          },
          "nextToken": {
            "type": "string",
            "description": "Token for next page of results",
            "example": "eyJsYXN0S2V5Ijp7InByb2R1Y3RJZCI6InByb2QtMTIzIn19"
          }
        }
      },
      "CreateProductRequest": {
        "type": "object",
        "required": ["name", "price", "currency", "category", "inventory"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Wireless Headphones"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "example": "Premium noise-cancelling wireless headphones"
          },
          "price": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "example": 199.99
          },
          "currency": {
            "type": "string",
            "enum": ["USD", "EUR", "GBP"],
            "example": "USD"
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "subCategory": {
            "type": "string",
            "example": "Audio"
          },
          "brand": {
            "type": "string",
            "example": "TechBrand"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "example": ["https://cdn.example.com/img1.jpg"]
          },
          "inventory": {
            "type": "integer",
            "minimum": 0,
            "example": 50
          },
          "sku": {
            "type": "string",
            "example": "WH-123-BLK"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "color": "Black",
              "weight": "250g"
            }
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive"],
            "default": "active",
            "example": "active"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Invalid product ID format"
          },
          "details": {
            "type": "object",
            "description": "Additional error details",
            "additionalProperties": true
          }
        }
      }
    }
  }
}